requirements.txt
python-dotenv
fastapi
uvicorn[standard]
python-multipart
twilio
supabase
requests
# AI dependencies - using compatible versions
openai>=1.0.0
langgraph>=0.0.20
langchain-core>=0.1.0